e-commerce - laravel 1.21

+++ database/migrations/2023_10_25_192146_create_products_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->longText('shortDescription')->nullable();
             $table->longText('longDescription')->nullable();
             $table->decimal('price', 10, 2);
-            $table->decimal('stock', 10, 2)->nullable();
+            $table->integer('stock')->nullable();
             $table->boolean('value');
             $table->timestamps();
             $table->softDeletes();
